+++ src/contributions/_index.ts
@@ -4,6 +4,7 @@ import registerCodeLensProviders from "./codelens/_index";
 import registerEditorProviders from "./editors/_index";
 import registerLanguageProviders from "./language/_index";
 import registerStatusBarItems from "./status-bars/_index";
+import registerCompletionProviders from "./completions/_index";
 
 export default function registerContributions(context: vscode.ExtensionContext) {
 	registerCommands();
@@ -11,4 +12,5 @@ export default function registerContributions(context: vscode.ExtensionContext) 
 	registerLanguageProviders(context);
 	registerEditorProviders(context);
 	registerStatusBarItems(context);
+	registerCompletionProviders(context);
 }