+++ src/contributions/completions/_index.ts
@@ -1,0 +1,6 @@
+import * as vscode from "vscode";
+import registerXmlCompletionProvider from "./xml-references";
+
+export default function registerCompletionProviders(context: vscode.ExtensionContext) {
+  registerXmlCompletionProvider(context);
+}